Lots of changes here.
Download directory is now user specified and logs are now stored in a Log dir. There is a downloading animation that plays during file retrieval, when download finishes a jingle plays. Now using threaded tasks, will be implementing more tasks to use separate threads for efficiency and to perceived freezing.